refactor(rolls): extract bank-roll module
refs #559
